pyproject.toml
[project]
name = "ai-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "google-genai==1.12.1",
    "llama-index>=0.14.15",
    "llama-index-llms-ollama>=0.9.1",
    "llama-index-tools-mcp>=0.4.7",
    "mcp[cli]>=1.26.0",
    "nest-asyncio>=1.6.0",
    "python-dotenv==1.1.0",
    "pytz>=2025.2",
]
